adds Drops to Languages
I really enjoy Drops, more than Duolingo, so I thought other people might too. 

FYI: Your Contributing.md is gone. You probably want to either put it back, or erase the line at the top of the README that references it. I was going to create an issue to say this, but it looks like I don't have permission to do that.